fix: correct nine correctness bugs in core image/font code
- WriteImageFile: capture and return encoder errors (were discarded);
  stop closing the caller's file, matching its doc and ReadImageFile
- IDrawEnclosedArc: multiply radius in float64 instead of truncating
  the trig result to -1/0/1 before scaling
- DrawStringRotatedAngle: iterate over runes and index text[loop:loop+1];
  the old text[loop:loop] was always empty so it rendered nothing
- BDF parser: SLANT now sets font.slant (not font.weight); DWIDTH parses
  only the horizontal advance (was clobbering char.height via an arg/verb
  mismatch); removed a duplicated unreachable PIXEL_SIZE branch
- FillPolygon: draw the span for any intersection instead of returning a
  leftover debug error sentinel
- FloodFill: rewrite as an iterative scanline fill (no recursion/stack
  overflow) with proper bounds checks and a same-color early-out

+++ arc.go
@@ -51,8 +51,8 @@ func IDrawEnclosedArc(image *Image, gc GraphicsContext, x int, y int, r1 int, r2
 	lastx := 0
 	lasty := 0
 	for loop := 0; loop < N; loop++ {
-		myx := x + int(r1*int(math.Cos(a+float64(loop)*da)))
-		myy := y + int(r2*int(math.Sin(a+float64(loop)*da)))
+		myx := x + int(float64(r1)*math.Cos(a+float64(loop)*da))
+		myy := y + int(float64(r2)*math.Sin(a+float64(loop)*da))
 		if loop > 0 {
 			DrawLine(image, gc, lastx, lasty, myx, myy)
 		}
